resourcestore --> idstore, store --> filestore
1. filestore now follows convention like labelstore, idstore, typestore.
2. store package will now house ResourceStore, the overarching store which utilizes the sub stores.
